+++ src/components/HeatmapBox/HeatmapBox.tsx
@@ -1,11 +1,11 @@
-import { PropsWithChildren } from "react";
+import { ReactNode } from "react";
 import { Box } from "src/types";
 
-interface HeatmapBoxProps extends PropsWithChildren {
+interface HeatmapBoxProps {
   box: Box;
 }
 
-export function HeatmapBox({ children, box }: HeatmapBoxProps) {
+export function HeatmapBox({ box }: HeatmapBoxProps) {
   const boxClassNames = [
     "heatmap-tracker-box",
     box.name,
@@ -18,13 +18,22 @@ export function HeatmapBox({ children, box }: HeatmapBoxProps) {
       : "isEmpty",
   ];
 
+  const content =
+  box.content instanceof HTMLElement ? (
+    <span dangerouslySetInnerHTML={{ __html: box.content.outerHTML }} />
+  ) : (
+    (box.content as ReactNode)
+  );
+
   return (
     <div
       data-date={box.date}
       style={{ backgroundColor: box.backgroundColor }}
       className={`${boxClassNames.filter(Boolean).join(" ")}`}
+      // On Desktop it will show the date on hover.
+      aria-label={box.date}
     >
-      <span className="heatmap-tracker-content">{children}</span>
+      <span className="heatmap-tracker-content">{content}</span>
     </div>
   );
 }